% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_confidence_plot}
\alias{create_confidence_plot}
\title{Create Confidence Plot}
\usage{
create_confidence_plot(umap_data, point_size, point_alpha)
}
\description{
Create Confidence Plot
}
\keyword{internal}
