% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_predict_plots}
\alias{create_predict_plots}
\title{Create PREDICT Mode Plots}
\usage{
create_predict_plots(
  seurat_obj,
  plot_type,
  point_size,
  point_alpha,
  return_data
)
}
\description{
Create PREDICT Mode Plots
}
\keyword{internal}
