% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_predictions_plot}
\alias{create_predictions_plot}
\title{Create Predictions Plot}
\usage{
create_predictions_plot(umap_data, point_size, point_alpha)
}
\description{
Create Predictions Plot
}
\keyword{internal}
