% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Default value operator}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{Left-hand side value}

\item{rhs}{Right-hand side value (default)}
}
\value{
lhs if not NULL, otherwise rhs
}
\description{
Returns the left-hand side if it's not NULL, otherwise returns the right-hand side.
This is a common utility operator used throughout the package.
}
\keyword{internal}
