% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{perform_tsne}
\alias{perform_tsne}
\title{Perform t-SNE dimensionality reduction}
\usage{
perform_tsne(seurat_object, dims = 1:40, perplexity = 30, max_iter = 1000)
}
\arguments{
\item{seurat_object}{A Seurat object containing PCA results.}

\item{dims}{Integer vector specifying which PCA dimensions to use.
Default is 1:40.}

\item{perplexity}{Numeric value specifying the perplexity parameter.
Default is 30.}

\item{max_iter}{Integer specifying the maximum number of iterations.
Default is 1000.}
}
\value{
A Seurat object with t-SNE coordinates added to the metadata.
}
\description{
Performs t-Distributed Stochastic Neighbor Embedding (t-SNE) on a Seurat object
using PCA results as input.
}
\examples{
seurat_object <- perform_tsne(seurat_object, dims = 1:40)
}
