% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{process_metadata}
\alias{process_metadata}
\title{Process metadata for Seurat object}
\usage{
process_metadata(seurat_object, metadata, barcodes)
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{metadata}{Data frame containing metadata}

\item{barcodes}{Data frame containing barcodes}
}
\value{
A Seurat object with processed metadata
}
\description{
Internal helper function to process metadata and barcodes for a Seurat object
}
\keyword{internal}
