% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_predict_mode}
\alias{run_predict_mode}
\title{Helper function for predict mode}
\usage{
run_predict_mode(
  seurat_object,
  fingerprint_file,
  output_dir,
  progress,
  pb,
  verbose
)
}
\description{
Helper function for predict mode
}
\keyword{internal}
